Test for first class callable issue
