+++ tests/Type/Doctrine/data/QueryResult/queryBuilderGetQuery.php
@@ -233,39 +233,4 @@ class QueryBuilderGetQuery
 		assertType('mixed', $result);
 	}
 
-	public function testQueryTypeIsInferredOnAcrossMethods(EntityManagerInterface $em): void
-	{
-		$query = $this->getQueryBuilder($em)->getQuery();
-		$branchingQuery = $this->getBranchingQueryBuilder($em)->getQuery();
-
-		assertType('Doctrine\ORM\Query<null, QueryResult\Entities\Many>', $query);
-		assertType('Doctrine\ORM\Query<null, QueryResult\Entities\Many>', $branchingQuery);
-	}
-
-	public function testQueryTypeIsInferredOnAcrossMethodsEvenWhenVariableAssignmentIsUsed(EntityManagerInterface $em): void
-	{
-		$queryBuilder = $this->getQueryBuilder($em);
-
-		assertType('Doctrine\ORM\Query<null, QueryResult\Entities\Many>', $queryBuilder->getQuery());
-	}
-
-	private function getQueryBuilder(EntityManagerInterface $em): QueryBuilder
-	{
-		return $em->createQueryBuilder()
-			->select('m')
-			->from(Many::class, 'm');
-	}
-
-	private function getBranchingQueryBuilder(EntityManagerInterface $em): QueryBuilder
-	{
-		$queryBuilder = $em->createQueryBuilder()
-			->select('m')
-			->from(Many::class, 'm');
-
-		if (random_int(0, 1) === 1) {
-			$queryBuilder->andWhere('m.intColumn = 1');
-		}
-
-		return $queryBuilder;
-	}
 }
